refactor: extract support control cli commands
